Finall update to compression code to make swapping in/out different compression libs easier (hopefully)
