+++ 3rdParty/LightPcapNg/LightPcapNg/include/light_pcapng.h
@@ -82,7 +82,6 @@ light_pcapng light_read_from_memory(const uint32_t *memory, size_t size);
 uint32_t *light_pcapng_to_memory(const light_pcapng pcapng, size_t *size);
 
 size_t light_pcapng_to_file_stream(const light_pcapng pcapng, light_file file);
-//size_t light_pcapng_to_compressed_file_stream(const light_pcapng pcapng, light_file file,const light_compression compression_context);
 
 int light_pcapng_to_file(const char *file_name, const light_pcapng pcapng);
 int light_pcapng_to_compressed_file(const char *file_name, const light_pcapng pcapng, int compression_level);
